refactor(phase0): extract pure shared utilities with compatibility wrappers
